chore(python-pro): organize, optimize, document the engineering core
Five-pass cleanup driven by the python-pro skill workflow. Touches
the engineering layers (topology / optimize / design / physics /
models) — UI is intentionally left alone (separate ux-ui-pro task).

Pass 1 — lint cleanup
  - engine.py: drop unused ``P_total`` unpack (RUF059); remove
    redundant ``int()`` cast on already-int ``N`` (RUF046).
  - interleaved_boost_pfc.py: same RUF046 on ``math.floor`` result.
  - buck_ccm.py: drop dead ``# noqa: D401`` (RUF100).

Pass 2 — extract shared DC-input Vin helpers
  - New ``topology/_dc_input.py``: single source of truth for the
    Vin_dc_min / Vin_dc_max / Vin_dc / legacy Vrms fallback chain.
  - buck_ccm.py + flyback.py: replace 99 %-identical private
    ``_vin_min/max/nom`` copies with thin aliases delegating to
    ``_dc_input``. Future PSFB output choke gets the same semantics
    for free instead of forking a third copy.

Pass 3 — replace topology if/elif chains with dispatch tables
  - feasibility.py: extract ``_REQUIRED_L_BY_TOPOLOGY`` and
    ``_PEAK_CURRENT_BY_TOPOLOGY`` dicts (typed via ``Callable``
    aliases). Adding a new topology now needs one dict insert
    instead of editing 6 elif chains across 2 functions; falls
    back to passive-choke for unmapped IDs instead of silently
    using the boost branch.

Pass 4 — module docstrings on the model layer
  - models/material.py, core.py, spec.py, wire.py, result.py:
    each gets a top-level docstring partitioning fields by
    topology family (universal / AC input / DC input / switching
    / line reactor / interleaved / flyback) so a reader doesn't
    have to grep call sites to figure out which fields belong
    to which converter type.

Pass 5 — type the Numba kernel factories
  - boost_ccm.py, core_loss.py, fused_kernel.py, engine.py: every
    ``_build_*_kernel()`` now annotates
    ``Optional[Callable[..., …]]`` and the inner ``_kernel`` body
    types each scalar / array param. Numba ships no .pyi stubs,
    so the wrapped imports get ``# type: ignore[import-untyped]``
    where mypy would otherwise complain.
  - engine._try_fused_thermal: full keyword-arg signature typed
    with the ``(T_final, P_total, P_cu_dc, P_cu_ac, P_core_line,
    P_core_ripple, converged) | None`` return contract spelled
    out instead of left as bare ``tuple``.
  - engine.py: drop dead ``if material.steinmetz is None``
    branch — ``Material.steinmetz`` is a required field, mypy
    correctly flags the check as unreachable.

Verification
  - ruff check src/: clean
  - ruff format src/: clean
  - mypy on topology+models+optimize+design (48 files): 0 errors
    (was 25 before pass 5)
  - pytest regression suite (103 tests across topology, engine,
    feasibility, core_loss, realistic_waveforms, report,
    topology_picker, cascade_protocol): green
  - tests/test_performance_kernels.py + tests/test_schematic_widget.py:
    auto-formatted by ruff (collapsed multi-line f-strings); no
    behavioral change.
  - uv.lock: package version bump 0.1.0 → 0.4.0 (resync from
    earlier release work, not a python-pro change but caught in
    same staging area).

Files changed: 14 modified, 1 new (``_dc_input.py``).

+++ src/pfc_inductor/topology/interleaved_boost_pfc.py
@@ -153,7 +153,9 @@ def ripple_cancellation_factor(D: float, N: int) -> float:
         return 0.0
     # Pick the cancellation cell. ``k`` in the Hwu-Yau paper is
     # the integer part of N·D + 1 (so D ∈ (0, 1/N] uses k=1, etc.).
-    k = int(math.floor(N * D)) + 1
+    # ``math.floor`` already returns an int — the explicit ``int()``
+    # cast was redundant.
+    k = math.floor(N * D) + 1
     if k > N:
         return 0.0
     num = (1.0 - k * D + (k - 1)) * (k * D - (k - 1))
